Create a SearchSuggestion component to refactoring
